feat: organize and create components to archive call
