+++ phone-test/src/pages/Calls/CallDetails/components/CallNote.tsx
@@ -1,0 +1,15 @@
+import { FC } from 'react';
+import { Spacer, NotesOutlined } from '@aircall/tractor';
+
+interface CallNoteProps {
+  content: string;
+}
+
+export const CallNote: FC<CallNoteProps> = ({ content }) => {
+  return (
+    <Spacer space="s" direction="horizontal" alignItems="center" justifyItems="center">
+      <NotesOutlined />
+      <span>{content}</span>
+    </Spacer>
+  );
+};
