+++ phone-test/src/pages/Calls/CallsList/CallsList.tsx
@@ -1,0 +1,68 @@
+import { useQuery } from '@apollo/client';
+import styled from 'styled-components';
+import { PAGINATED_CALLS } from '../../../gql/queries';
+import { Typography, Spacer, Pagination } from '@aircall/tractor';
+import { CallItem } from './components/CallItem';
+import { useNavigate, useSearchParams } from 'react-router-dom';
+
+export const PaginationWrapper = styled.div`
+  > div {
+    width: inherit;
+    margin-top: 20px;
+    display: flex;
+    justify-content: center;
+  }
+`;
+
+const CALLS_PER_PAGE = 5;
+
+export const CallsListPage = () => {
+  const [search] = useSearchParams();
+  const navigate = useNavigate();
+  const pageQueryParams = search.get('page');
+  const activePage = !!pageQueryParams ? parseInt(pageQueryParams) : 1;
+  const { loading, error, data } = useQuery(PAGINATED_CALLS, {
+    variables: {
+      offset: (activePage - 1) * CALLS_PER_PAGE,
+      limit: CALLS_PER_PAGE
+    }
+    // onCompleted: () => handleRefreshToken(),
+  });
+
+  if (loading) return <p>Loading calls...</p>;
+  if (error) return <p>ERROR</p>;
+  if (!data) return <p>Not found</p>;
+
+  const { totalCount, nodes: calls } = data.paginatedCalls;
+
+  const handlePageChange = (page: number) => {
+    navigate(`/calls/?page=${page}`);
+  };
+
+  return (
+    <>
+      <Typography variant="displayM" textAlign="center" py={3}>
+        Calls History
+      </Typography>
+      <Spacer space={3} direction="vertical">
+        {calls.map((call: Call) => {
+          return (
+            <div key={call.id}>
+              <CallItem call={call} />
+            </div>
+          );
+        })}
+      </Spacer>
+      {totalCount && (
+        <PaginationWrapper>
+          <Pagination
+            activePage={activePage}
+            pageSize={CALLS_PER_PAGE}
+            onPageChange={handlePageChange}
+            recordsTotalCount={totalCount}
+          />
+        </PaginationWrapper>
+      )}
+    </>
+  );
+};
